requirements.txt
discord.py==2.5.2
protobuf==6.30.2
python-dotenv==1.1.0
google-generativeai==0.8.4
